refactor(base-ui): rename components & remove unused code
