+++ packages/base-ui/package.json
@@ -52,6 +52,7 @@
         "@univerjs/design": "workspace:*",
         "@univerjs/icons": "^0.1.3",
         "@wendellhu/redi": "^0.12.10",
+        "clsx": "^2.0.0",
         "localforage": "^1.10.0",
         "rc-menu": "^9.12.2",
         "rc-notification": "^5.3.0",
