--- packages/base-ui/src/Components/Input/Style/index.module.less
@@ -1,5 +1,0 @@
-// @import './mixin.less';
-
-// .input {
-//     .input();
-// }
