--- packages/base-ui/src/Components/Input/Style/mixin.less
@@ -1,81 +1,0 @@
-// @import '../../../Basics/CSS/Skin/default.module.less';
-
-// .active() {
-//     outline: 0;
-//     box-shadow: var(--box-shadow-base);
-// }
-
-// .hover(@color: var(--primary-color-hover)) {
-//     border-color: @color;
-// }
-
-// .disabled() {
-//     color: var(--grey-200);
-//     background-color: var(--bg-color-secondary);
-//     border-color: var(--grey-200);
-//     box-shadow: none;
-//     cursor: not-allowed;
-//     opacity: 1;
-
-//     &:hover {
-//         .hover(var(--grey-200));
-//     }
-// }
-
-// .input() {
-//     position: relative;
-//     display: inline-block;
-//     width: 100%;
-//     height: 100%;
-//     box-sizing: border-box;
-//     padding: 10px 20px;
-//     color: var(--text-color);
-//     font-size: var(--font-size-base);
-//     background-color: var(--bg-color-secondary);
-//     border: 1px solid var(--grey-200);
-//     border-radius: var(--border-radius-base);
-//     transition: all 0.3s;
-
-//     &:hover {
-//         .hover();
-//     }
-
-//     &:focus,
-//     &-focused {
-//         .active();
-//     }
-
-//     &[disabled] {
-//         .disabled();
-//     }
-
-//     &-disabled {
-//         .disabled();
-//     }
-
-//     &-borderless {
-//         border: 1px solid transparent;
-//         box-shadow: none;
-
-//         &:hover {
-//             border-color: transparent;
-//         }
-
-//         &:focus,
-//         &-focused {
-//             border-color: var(--primary-color-hover);
-//             .active();
-//         }
-//     }
-
-//     // Remove Arrows/Spinners
-//     &::-webkit-outer-spin-button,
-//     &::-webkit-inner-spin-button{
-//         -webkit-appearance: none;
-//         margin: 0;
-//     }
-
-//     &[type=number] {
-//         -moz-appearance: textfield;
-//     }
-// }
