--- packages/base-ui/src/Components/Input/index.ts
@@ -1,1 +1,0 @@
-export * from './Input';
