--- packages/base-ui/src/Components/Select/index.ts
@@ -1,1 +1,0 @@
-export * from './Select';
