--- packages/base-ui/src/Components/Tabs/index.module.less
@@ -1,86 +1,0 @@
-.tabs {
-    overflow: hidden;
-
-    ::-webkit-scrollbar {
-        display: none;
-    }
-
-    .tabs-nav {
-        width: 100%;
-
-        .tabs-nav-wrap {
-            width: 100%;
-            position: relative;
-            display: flex;
-            flex: auto;
-            align-self: stretch;
-            overflow: hidden;
-            white-space: nowrap;
-
-            .tabs-nav-list {
-                display: inline-block;
-                overflow-y: hidden;
-                overflow-x: auto;
-                position: relative;
-                white-space: nowrap;
-                scroll-behavior: smooth;
-                transition: scroll-behavior 0.3s ease-in-out;
-            }
-        }
-    }
-
-    .tabs-content {
-        display: flex;
-        width: 100%;
-
-        .tabs-panel {
-            display: none;
-            width: 100%;
-            outline: none;
-
-            &-active {
-                display: block;
-            }
-        }
-    }
-}
-
-.tabs-tab {
-    display: inline-block;
-    position: relative;
-    user-select: none;
-    padding: 0.5rem 1em;
-    font-size: var(--font-size);
-    background: transparent;
-    border: 0;
-    outline: none;
-    cursor: pointer;
-
-    &:hover,
-    &-active {
-        color: var(--primary-color);
-    }
-}
-
-.tabs.tabs-card {
-    .tabs-nav {
-        width: 100%;
-        .tabs-nav-wrap {
-            width: 100%;
-            .tabs-nav-list {
-                .tabs-tab {
-                    margin: 0;
-                    // background: @bg-color;
-                    // border: 1px solid @border-color;
-
-                    &-active {
-                        color: var(--primary-color);
-                        background: var(--bg);
-                        border-bottom: 1px solid var(--bg);
-                    }
-                }
-            }
-        }
-    }
-}
-
