--- packages/base-ui/src/Components/Tabs/index.ts
@@ -1,1 +1,0 @@
-export * from './Tabs';
