--- packages/base-ui/src/Components/index.ts
@@ -1,6 +1,0 @@
-export * from './CustomLabel';
-export * from './Input';
-export * from './Menu';
-export * from './Select';
-export * from './Slider';
-export * from './Tabs';
