+++ packages/base-ui/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './custom-label';
+export * from './menu';
+export * from './slider';
