+++ packages/base-ui/src/components/notification/index.module.less
