+++ packages/base-ui/src/components/slider/Slider.stories.tsx
