+++ packages/base-ui/src/components/slider/__test__/index.spec.tsx
