+++ packages/base-ui/src/components/slider/index.module.less
