+++ packages/base-ui/src/components/slider/index.ts
