+++ packages/base-ui/src/views/components/DocBars/Select/index.module.less
@@ -14,10 +14,6 @@
         border-radius: var(--border-radius-base);
     }
 
-    .select-button {
-
-    }
-
     .select-single {
         text-overflow: ellipsis;
         white-space: nowrap;
