+++ packages/design/src/components/input-number/index.module.less
@@ -21,11 +21,11 @@
         }
     }
 
-    &-out-of-range {
-        .input-number-input {
-            color: rgb(var(--error-color));
-        }
-    }
+    // &-out-of-range {
+    //     .input-number-input {
+    //         color: rgb(var(--error-color));
+    //     }
+    // }
 
     &-handler {
         text-align: center;
