+++ packages/design/src/components/input/index.module.less
@@ -7,9 +7,9 @@
         background-color: rgb(var(--primary-color));
     }
 
-    &-out-of-range {
-        color: rgb(var(--error-color));
-    }
+    // &-out-of-range {
+    //     color: rgb(var(--error-color));
+    // }
 
     &-affix-wrapper {
         border: 1px solid rgb(var(--border-color));
