+++ packages/sheets-plugin-numfmt/src/View/UI/FormatItem.tsx
@@ -1,8 +1,6 @@
-import { joinClassNames } from '@univerjs/base-ui';
+// import { joinClassNames } from '@univerjs/base-ui';
 import { LocaleService } from '@univerjs/core';
 import { useDependency } from '@wendellhu/redi/react-bindings';
-
-import styles from './FormatItem.module.less';
 
 export interface BaseFormatItemProps {
     selected?: boolean;
@@ -21,10 +19,11 @@ export function FormatItem(props: BaseFormatItemProps): JSX.Element {
     const localeService = useDependency(LocaleService);
 
     return (
-        <div className={joinClassNames(styles.formatItem, disabled ? styles.selectDisabledItem : '')}>
-            {selected && <span className={styles.formatItemSelected}>{/* <Icon.CorrectIcon /> */}</span>}
-            <span className={styles.formatItemContent}>{localeService.t(labelText as string) || labelText}</span>
-            {suffix && <span className={styles.formatItemSuffix}>{localeService.t(suffix as string) || suffix}</span>}
-        </div>
+        // <div className={joinClassNames(styles.formatItem, disabled ? styles.selectDisabledItem : '')}>
+        //     {selected && <span className={styles.formatItemSelected}>{/* <Icon.CorrectIcon /> */}</span>}
+        //     <span className={styles.formatItemContent}>{localeService.t(labelText as string) || labelText}</span>
+        //     {suffix && <span className={styles.formatItemSuffix}>{localeService.t(suffix as string) || suffix}</span>}
+        // </div>
+        <></>
     );
 }
